part1a: misc fixes
TerrainOverlay: fix warnings
PathfindEngine.cpp: correct float type, mark as noncopyable
TriggerManager: explicitly construct as CStrW

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5516 3db68df2-c116-0410-a063-a993310a9797

+++ source/renderer/TerrainOverlay.cpp
@@ -95,7 +95,7 @@ void TerrainOverlay::RenderEntityEdges()
 
 	
 	
-	for(int i =0 ; i < results.size(); i++)
+	for(size_t i =0 ; i < results.size(); i++)
 	{
 		glBegin(GL_LINE_LOOP);
